requirements.txt
torch
transformers
accelerate
ray[default]
fastapi
uvicorn
grpcio